+++ linkGrabber/tests/test_scrape.py
@@ -1,4 +1,5 @@
 """ Unit test Links functionality"""
+import os.path
 import unittest
 import requests
 import vcr
@@ -7,6 +8,8 @@ import bs4
 from linkGrabber import Links
 from linkGrabber.tests import test_data as td
 
+BASE_DIR = os.path.dirname(os.path.realpath(__file__))
+
 class TestScrape(unittest.TestCase):
     """ A set of unit tests for Links """
     def setUp(self):
@@ -14,10 +17,10 @@ class TestScrape(unittest.TestCase):
         self.url = "http://www.google.com"
         self.bad_url = "www.google.com"
         # grab some example html pages to test
-        with vcr.use_cassette('fixtures/vcr_cassettes/first_test.yaml'):
+        with vcr.use_cassette(os.path.join(BASE_DIR, 'fixtures', 'vcr_cassettes', 'first_test.yaml')):
             td.pages[0]['text'] = requests.get(td.pages[0]['url']).text
 
-        with vcr.use_cassette('fixtures/vcr_cassettes/second_test.yaml'):
+        with vcr.use_cassette(os.path.join(BASE_DIR, 'fixtures', 'vcr_cassettes', 'second_test.yaml')):
             td.pages[1]['text'] = requests.get(td.pages[1]['url']).text
 
     def test_url(self):
@@ -28,12 +31,6 @@ class TestScrape(unittest.TestCase):
         """ Getting the web page yields correct response"""
         seek = Links(self.url)
         self.assertIsInstance(seek._soup, bs4.BeautifulSoup)
-
-    def test_find_bad_filter_param(self):
-        """ Bad filters parameter inputs """
-        seek = Links(self.url)
-        self.assertRaises(Exception, seek.find, filters=25)
-        self.assertRaises(Exception, seek.find, filters=['href', 'style'])
 
     def test_find_limit_param(self):
         """ How does find() handle the limit property """
@@ -82,6 +79,12 @@ class TestScrape(unittest.TestCase):
             self.assertEqual(len(actual_list), len(page['limit_sort_href']))
             for i, link in enumerate(actual_list):
                 self.assertDictEqual(link, page['limit_sort_href'][i])
+
+    def test_find_exclude(self):
+        """ Determine if excluding links removes the links """
+
+    def test_find_duplicates(self):
+        """ Determine if removing duplicates works """
 
 if __name__ == '__main__':
     unittest.main(verbosity=2)